disk streamer absorbed into loader class
image object is created in disk reader and image file is set, then passed as normal to stream cam and subsequently to loader class. loader class now does a check to see if image is in memory, if not it loads it then proceeds as normal.

blur detection removed via commenting due to inconsistent results.