composite manager add new comp
add new component is now delayed until image index is appropriate. it takes far less time to realize that a new component needs to be created than to allow a frame to become part of an existing component. this can result in spawning a new component for index 100 even though only up to index 25 has been added to the first component. this delay ensures indexes 26-99 are composited before a new component spawned by image 100 is added to the screen.