image pyramid processing entirely on gpu now
except two small things: polymask output is a fillConvexPoly implementation that apparently is difficult to vectorize and is potentially not more efficient when done on the gpu. This should still probably be done on the gpu since now we have to upload it at every iteration. nbd for now though

and getTiles, which is a request made by the front end. This currently results in a backend gpumat.download (first copy) and then a Mat.data to juce.bitmap (second copy). a more efficient implementation would be to do a single cudaMemCpy from the gpumat to the juce.bitmap. However, I am holding off on this for now because this would mean that the front end (which is supposed to be interchangeable) would have to anticipate a different type of object depending on if the code is compiled w cuda support or not. Im leaving it how it is for now since this has no impact on the compositing performance, only on the front end updating the display.